Remove all NaClHandlePass*() calls on the NaCl side
The handle_pass library is no longer needed because it is replaced by
Chromium's BrokerDuplicateHandle() function.

I can't remove handle_pass just yet because there are still calls to
it on the Chromium side.

Chromium will be changed not to return a process handle to
sel_ldr_launcher_chrome.cc, so deal with child_process_ being NULL.
Remove the child_process() method because no-one is using it, and
because it won't be useful in the Chromium Windows sandbox.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2719
TEST=run_ppapi_ppb_core_browser_test in Chromium on Windows

Review URL: https://chromiumcodereview.appspot.com/10035007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@8290 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: a2105f6682ce988f3f43668a71b7b0135be77b13

+++ src/trusted/nonnacl_util/win/sel_ldr_launcher_win.cc
@@ -140,10 +140,11 @@ bool SelLdrLauncher::StartViaCommandLine(
 }
 
 bool SelLdrLauncher::KillChildProcess() {
-    return 0 != TerminateProcess(child_process_, 9);
-    // 9 is the exit code for the child_process_.  The value is actually not
-    // material, since (currently) the launcher does not collect/report
-    // it.
+  if (kInvalidHandle == child_process_)
+    return false;
+  return 0 != TerminateProcess(child_process_, 9);
+  // 9 is the exit code for the child_process_.  The value is actually not
+  // material, since (currently) the launcher does not collect/report it.
 }
 
 void PluginSelLdrLocator::GetDirectory(char* buffer, size_t len) {
